Componentize EnemyMoves
